requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Mail==0.9.1
email-validator==1.3.1
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1